🩹 fix : 팀 외래키 추가

+++ src/main/java/jira6/fate/domain/card/service/CardService.java
@@ -1,47 +1,66 @@
 package jira6.fate.domain.card.service;
 
-import static jira6.fate.global.exception.ErrorCode.COLUMN_NOT_FOUND;
-
-import jakarta.transaction.Transactional;
 import jira6.fate.domain.card.dto.CardCreateRequestDto;
+import jira6.fate.domain.card.dto.CardDetailResponseDto;
 import jira6.fate.domain.card.dto.CardUpdateRequestDto;
 import jira6.fate.domain.card.entity.Card;
 import jira6.fate.domain.card.repository.CardRepository;
 import jira6.fate.domain.user.entity.User;
+import jira6.fate.domain.user.repository.UserRepository;
 import jira6.fate.global.exception.CustomException;
 import jira6.fate.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.annotations.Columns;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class CardService {
 
     private final CardRepository cardRepository;
-    private final ColumnRepository columnRepository;
+    private final ColumnsRepository columnsRepository;
+    private final TeamRepository teamRepository;
 
     @Transactional
     public void createCard(Long columnId, CardCreateRequestDto requestDto, User user) {
-        Columns column = existsByColumn(columnId);
+        Columns columns = findColumn(columnId);
+        Team team = findTeam(requestDto.getTeamId());
 
         Card card = Card.builder()
             .cardTitle(requestDto.getCardTitle())
             .cardContents(requestDto.getCardContents())
             .cardOrder(requestDto.getCardOrder())
-            .managerName(requestDto.getManagerName())
             .deadlineAt(requestDto.getDeadlineAt())
-            .column(columns)
+            .columns(columns)
             .user(user)
+            .team(team)
             .build();
 
         cardRepository.save(card);
     }
 
+    @Transactional(readOnly = true)
+    public CardDetailResponseDto getCard(Long columnId, Long cardId) {
+        Columns columns = findColumn(columnId);
+        Card card = findCard(cardId);
+
+        String managerName = card.getTeam().getUser().getUserName();
+
+        return CardDetailResponseDto.builder()
+            .columnName(card.getCardTitle())
+            .cardContents(card.getCardContents())
+            .managerName(managerName)
+            .deadlineAt(card.getDeadlineAt())
+            .columnName(columns.getColumnName())
+            .build();
+    }
+
     @Transactional
     public void updateCard(Long columnId, Long cardId, CardUpdateRequestDto requestDto, User user) {
-        Columns column = findColumn(columnId);
+        Columns columns = findColumn(columnId);
         Card card = findCard(cardId);
+        Team team = findTeam(requestDto.getTeamId());
 
         String cardCreatorName = card.getUser().getUserName();
         String currentUserName = user.getUserName();
@@ -53,9 +72,9 @@ public class CardService {
         card.update(
             requestDto.getCardTitle(),
             requestDto.getCardContents(),
-            requestDto.getManagerName(),
             requestDto.getDeadlineAt(),
-            column
+            columns,
+            team
         );
     }
 
@@ -63,6 +82,13 @@ public class CardService {
     public void deleteCard(Long columnId, Long cardId, User user) {
         findColumn(columnId);
         Card card = findCard(cardId);
+
+        String cardCreatorName = card.getUser().getUserName();
+        String currentUserName = user.getUserName();
+
+        if (!checkCardCreator(cardCreatorName, currentUserName)) {
+            throw new CustomException(ErrorCode.NOT_UNAUTHORIZED);
+        }
 
         cardRepository.delete(card);
     }
@@ -76,6 +102,12 @@ public class CardService {
     public Card findCard(Long cardId) {
         return cardRepository.findById(cardId).orElseThrow(
             () -> new CustomException(ErrorCode.CARD_NOT_FOUND)
+        );
+    }
+
+    public Team findTeam(Long teamId) {
+        return teamRepository.findById(teamId).orElseThrow(
+            () -> new CustomException(ErrorCode.USER_NOT_TEAM)
         );
     }
